Extract saveRecording logic
