+++ src/main/java/io/cryostat/net/web/http/api/v1/TargetRecordingPatchSave.java
@@ -37,56 +37,27 @@
  */
 package io.cryostat.net.web.http.api.v1;
 
-import java.io.IOException;
-import java.io.InputStream;
-import java.net.URISyntaxException;
-import java.nio.file.Path;
-import java.time.temporal.ChronoUnit;
 import java.util.Map;
-import java.util.Optional;
 
 import javax.inject.Inject;
-import javax.inject.Named;
 
-import org.openjdk.jmc.rjmx.services.jfr.IRecordingDescriptor;
-
-import io.cryostat.MainModule;
-import io.cryostat.core.net.JFRConnection;
-import io.cryostat.core.sys.Clock;
-import io.cryostat.core.sys.FileSystem;
 import io.cryostat.messaging.notifications.NotificationFactory;
 import io.cryostat.net.ConnectionDescriptor;
-import io.cryostat.net.TargetConnectionManager;
 import io.cryostat.net.web.http.HttpMimeType;
-import io.cryostat.platform.PlatformClient;
-import io.cryostat.util.URIUtil;
+import io.cryostat.recordings.RecordingHelper;
 
 import io.vertx.ext.web.RoutingContext;
-import io.vertx.ext.web.handler.impl.HttpStatusException;
 
 class TargetRecordingPatchSave {
 
-    private final FileSystem fs;
-    private final Path recordingsPath;
-    private final TargetConnectionManager targetConnectionManager;
-    private final Clock clock;
-    private final PlatformClient platformClient;
     private final NotificationFactory notificationFactory;
+    private final RecordingHelper recordingCreationHelper;
     private static final String NOTIFICATION_CATEGORY = "RecordingArchived";
 
     @Inject
     TargetRecordingPatchSave(
-            FileSystem fs,
-            @Named(MainModule.RECORDINGS_PATH) Path recordingsPath,
-            TargetConnectionManager targetConnectionManager,
-            Clock clock,
-            PlatformClient platformClient,
-            NotificationFactory notificationFactory) {
-        this.fs = fs;
-        this.recordingsPath = recordingsPath;
-        this.targetConnectionManager = targetConnectionManager;
-        this.clock = clock;
-        this.platformClient = platformClient;
+            RecordingHelper recordingCreationHelper, NotificationFactory notificationFactory) {
+        this.recordingCreationHelper = recordingCreationHelper;
         this.notificationFactory = notificationFactory;
     }
 
@@ -94,27 +65,8 @@ class TargetRecordingPatchSave {
         String recordingName = ctx.pathParam("recordingName");
 
         String saveName =
-                targetConnectionManager.executeConnectedTask(
-                        connectionDescriptor,
-                        connection -> {
-                            Optional<IRecordingDescriptor> descriptor =
-                                    connection.getService().getAvailableRecordings().stream()
-                                            .filter(
-                                                    recording ->
-                                                            recording
-                                                                    .getName()
-                                                                    .equals(recordingName))
-                                            .findFirst();
-                            if (descriptor.isPresent()) {
-                                return saveRecording(connection, descriptor.get());
-                            } else {
-                                throw new HttpStatusException(
-                                        404,
-                                        String.format(
-                                                "Recording with name \"%s\" not found",
-                                                recordingName));
-                            }
-                        });
+                recordingCreationHelper.saveRecording(connectionDescriptor, recordingName);
+
         ctx.response().setStatusCode(200);
         ctx.response().end(saveName);
         notificationFactory
@@ -125,55 +77,5 @@ class TargetRecordingPatchSave {
                         Map.of("recording", saveName, "target", connectionDescriptor.getTargetId()))
                 .build()
                 .send();
-    }
-
-    private String saveRecording(JFRConnection connection, IRecordingDescriptor descriptor)
-            throws Exception {
-        String recordingName = descriptor.getName();
-        if (recordingName.endsWith(".jfr")) {
-            recordingName = recordingName.substring(0, recordingName.length() - 4);
-        }
-
-        // TODO: To avoid having to perform this lookup each time, we should implement
-        // something like a map from targetIds to corresponding ServiceRefs
-        String targetName =
-                platformClient.listDiscoverableServices().stream()
-                        .filter(
-                                serviceRef -> {
-                                    try {
-                                        return serviceRef
-                                                        .getServiceUri()
-                                                        .equals(
-                                                                URIUtil.convert(
-                                                                        connection.getJMXURL()))
-                                                && serviceRef.getAlias().isPresent();
-                                    } catch (URISyntaxException | IOException ioe) {
-                                        return false;
-                                    }
-                                })
-                        .map(s -> s.getAlias().get())
-                        .findFirst()
-                        .orElse(connection.getHost())
-                        .replaceAll("[\\._]+", "-");
-
-        String timestamp =
-                clock.now().truncatedTo(ChronoUnit.SECONDS).toString().replaceAll("[-:]+", "");
-        String destination = String.format("%s_%s_%s", targetName, recordingName, timestamp);
-        // TODO byte-sized rename limit is arbitrary. Probably plenty since recordings are also
-        // differentiated by second-resolution timestamp
-        byte count = 1;
-        while (fs.exists(recordingsPath.resolve(destination + ".jfr"))) {
-            destination =
-                    String.format("%s_%s_%s.%d", targetName, recordingName, timestamp, count++);
-            if (count == Byte.MAX_VALUE) {
-                throw new IOException(
-                        "Recording could not be saved. File already exists and rename attempts were exhausted.");
-            }
-        }
-        destination += ".jfr";
-        try (InputStream stream = connection.getService().openStream(descriptor, false)) {
-            fs.copy(stream, recordingsPath.resolve(destination));
-        }
-        return destination;
     }
 }
